Module malachite_nz::natural::arithmetic

Traits for arithmetic.

Modules

add

Addition of Naturals.

add_mul

Implementations of AddMul and AddMulAssign, traits for adding a number and the product of two other numbers.

binomial_coefficient

Implementations of BinomialCoefficient, a trait for computing the binomial coefficient of two numbers.

checked_sub

Implementations of CheckedSub, a trait for subtracting two numbers and checking whether the result is representable.

checked_sub_mul

Implementations of CheckedSubMul, a trait for subtracting the product of two numbers from another number, and checking whether the result is representable.

coprime_with

Implementations of CoprimeWith, a trait for determining whether two numbers are coprime.

div

Division of Naturals.

div_exact

Implementations of DivExact and DivExactAssign, traits for dividing two numbers when it’s known that the division is exact.

div_mod

Implementations of raits for simultaneously finding the quotient and remainder of two numbers, subject to various rounding rules.

div_round

Implementations of DivRound and DivExactAssign, traits for dividing two numbers according to a specified RoundingMode.

divisible_by

Implementations of DivisibleBy, a trait for determining whether one number is divisible by another.

divisible_by_power_of_2

Implementations of DivisibleByPowerOf2, a trait for determining whether a number is divisible by $2^k$.

eq_mod

Implementations of EqMod, a trait for determining whether one number is equal by another modulo a third.

eq_mod_power_of_2

Implementations of EqModPowerOf2, a trait for determining whether one number is equal to another modulo $2^k$.

factorial

Implementations of Factorial, DoubleFactorial, Multifactorial, and Subfactorial.

gcd

Implementations of Gcd and GcdAssign, traits for computing the GCD (greatest common divisor) of two numbers.

is_power_of_2

Implementations of IsPowerOf2, a trait for determining whether a number is an integer power of 2.

kronecker_symbol

Implementations of LegendreSymbol, JacobiSymbol, and KroneckerSymbol, traits for computing the Legendre, Jacobi, and Kronecker symbols of two numbers.

lcm

Implementations of Lcm, LcmAssign, and CheckedLcm, traits for computing the LCM (least common multiple) of two numbers.

log_base

Implementations of traits for taking the base-$b$ logarithm of a number.

log_base_2

Implementations of traits for taking the base-2 logarithm of a number.

log_base_power_of_2

Implementations of traits for taking the base-$2^k$ logarithm of a number.

mod_add

Implementations of ModAdd and ModAddAssign, traits for adding two numbers modulo another number.

mod_inverse

Implementations of ModInverse, a trait for finding the multiplicative inverse of a number modulo another number.

mod_is_reduced

Implementations of ModIsReduced, a trait for checking whether a number is reduced modulo another number.

mod_mul

Implementations of traits for multiplying two numbers modulo another number.

mod_neg

Implementations of ModNeg and ModNegAssign, traits for negating a number modulo another number.

mod_op

Implementations of traits for finding the remainder of two numbers, subject to various rounding rules.

mod_pow

Implementations of traits for raising a number to a power modulo another number.

mod_power_of_2

Implementations of traits for finding the remainder of a number divided by $2^k$, subject to various rounding rules.

mod_power_of_2_add

Implementations of ModPowerOf2Add and ModPowerOf2AddAssign, traits for adding two numbers modulo $2^k$.

mod_power_of_2_inverse

Implementations of ModPowerOf2Inverse, a trait for finding the multiplicative inverse of a number modulo $2^k$.

mod_power_of_2_is_reduced

Implementations of ModPowerOf2IsReduced, a trait for checking whether a number is reduced modulo $2^k$.

mod_power_of_2_mul

Implementations of ModPowerOf2Mul and ModPowerOf2MulAssign, traits for multiplying two numbers modulo $2^k$.

mod_power_of_2_neg

Implementations of ModPowerOf2Neg and ModPowerOf2NegAssign, traits for negating a number modulo $2^k$.

mod_power_of_2_pow

Implementations of ModPowerOf2Pow and ModPowerOf2PowAssign, traits for raising a number to a power modulo $2^k$.

mod_power_of_2_shl

Implementations of ModPowerOf2Shl and ModPowerOf2ShlAssign, traits for left-shifting a number modulo $2^k$.

mod_power_of_2_shr

Implementations of ModPowerOf2Shr and ModPowerOf2ShrAssign, traits for right-shifting a number modulo $2^k$.

mod_power_of_2_square

Implementations of ModPowerOf2Square](malachite_base::num::arithmetic::traits::ModPowerOf2Square) and ModPowerOf2SquareAssign, traits for squaring a number modulo $2^k$.

mod_power_of_2_sub

Implementations of ModPowerOf2Sub and ModPowerOf2SubAssign, traits for subtracting one number by another modulo $2^k$.

mod_shl

Implementations of ModShl and ModShlAssign, traits for left-shifting a number modulo another number.

mod_shr

Implementations of ModShr and ModShrAssign, traits for right-shifting a number modulo another number.

mod_square

Implementations of traits for squaring a number modulo another number.

mod_sub

Implementations of ModSub and ModSubAssign, traits for subtracting two numbers modulo another number.

mul

Multiplication of Naturals.

neg

Negation of a Natural, returning an Integer.

next_power_of_2

Implementations of NextPowerOf2 and NextPowerOf2Assign, traits for getting the next-highest power of 2.

parity

Implementations of Parity, a trait for determining whether a number is even or odd.

pow

Implementations of Pow and PowAssign, traits for raising a number to a power.

power_of_2

Implementations of PowerOf2, a trait for computing a power of 2.

primorial

An implementation of Primorial, a trait for computing the primorial of a number.

root

Implementations of traits for taking the $n$th root of a number.

round_to_multiple

Implementations of RoundToMultiple and RoundToMultipleAssign, traits for rounding a number to a multiple of another number.

round_to_multiple_of_power_of_2

Implementations of RoundToMultipleOfPowerOf2 and RoundToMultipleOfPowerOf2Assign, traits for rounding a number to a multiple of a power of 2.

saturating_sub

Implementations of SaturatingSub and SaturatingSubAssign, traits for subtracting two numbers and saturating at numeric bounds instead of overflowing.

saturating_sub_mul

Implementations of SaturatingSubMul and SaturatingSubMulAssign, traits for subtracting a number by the product of two numbers and saturating at numeric bounds instead of overflowing.

shl

Left-shifting a Natural (multiplying it by a power of 2).

shl_round

Implementations of ShlRound and ShlRoundAssign, traits for multiplying a number by a power of 2 and rounding according to a specified RoundingMode.

shr

Right-shifting a Natural (dividing it by a power of 2).

shr_round

Implementations of ShrRound and ShrRoundAssign, traits for dividing a number by a power of 2 and rounding according to a specified RoundingMode.

sign

Implementations of Sign, a trait for determining the sign of a number.

sqrt

Implementations of traits for taking the square root of a number.

square

Implementations of Square and SquareAssign, traits for squaring a number.

sub

Subtraction of Naturals.

sub_mul

Implementations of SubMul and SubMulAssign, traits for subtracting the product of two numbers from a number.