Module malachite_nz::natural::conversion

Traits for converting to and from Naturals, converting to and from strings, and extracting digits.

Modules

digits

Implementations of traits for working with the digits of Naturals.

from_bool

An implementation of the From trait for converting a bool to a Natural.

from_limbs

Functions for constructing a Natural from Limbs.

from_primitive_float

Implementations of traits for converting a primitive float to a Natural.

from_primitive_int

Implementations of traits for converting a primitive integer to a Natural.

is_integer

An implementation of IsInteger, a trait for determining whether a number is an integer.

limb_count

A function for counting a Natural’s Limbs.

mantissa_and_exponent

Implementations of traits for converting numbers to and from mantissa-and-exponent representations.

primitive_float_from_natural

Implementations of traits for converting a Natural to a primitive float.

primitive_int_from_natural

Implementations of traits for converting a Natural to a primitive integer.

serde

Implementations of traits for serialization and deserialization using serde.

string

Implementations of traits for converting Naturals to and from Strings.

to_limbs

Functions for extracting Limbs from a Natural.