Module malachite_nz::natural::exhaustive

Iterators that generate Naturals without repetition.

Structs

ExhaustiveNaturalRange

Generates all Naturals in a finite interval.

ExhaustiveNaturalRangeToInfinity

Generates all Naturals greater than or equal to some Natural, in ascending order.

Functions

exhaustive_natural_inclusive_range

Generates all Naturals in the closed interval $[a, b]$, in ascending order.

exhaustive_natural_range

Generates all Naturals in the half-open interval $[a, b)$, in ascending order.

exhaustive_natural_range_to_infinity

Generates all Naturals greater than or equal to some number $a$, in ascending order.

exhaustive_naturals

Generates all Naturals in ascending order.

exhaustive_positive_naturals

Generates all positive Naturals in ascending order.