Module malachite_nz::natural::logic

Traits for logic and bit manipulation.

Modules

and

Bitwise and of Naturals.

bit_access

An implementation of BitAccess, a trait for getting and setting individual bits of a number.

bit_block_access

An implementation of BitBlockAccess, a trait for getting and setting adjacent blocks of bits in a number.

bit_convertible

An implementation of BitConvertible, a trait for extracting all bits from a number or constructing a number from bits.

bit_iterable

An implementation of BitIterable, a trait for producing a double-ended iterator over a number’s bits.

bit_scan

An implementation of BitScan, a trait for finding the next true or false bit in a number after a provided index.

count_ones

An implementation of CountOnes, a trait for counting the number of ones in the binary representation of a number.

hamming_distance

An implementation of HammingDistance, a trait for computing the Hamming distance between two numbers.

low_mask

An implementation of LowMask, a trait for generating a low bit mask (a number in which only the $k$ least-significant bits are 1).

not

Bitwise negation of Naturals.

or

Bitwise or of Naturals.

significant_bits

An implementation of SignificantBits, a trait for determining how many significant bits a number has.

trailing_zeros

An implementation of TrailingZeros, a trait for determining the number of zeros that a number ends with when written in binary.

xor

Bitwise xor of Naturals.