Module malachite_nz::integer::conversion::string

Implementations of traits for converting Integers to and from Strings.

Modules

from_sci_string

Implementations of FromSciString. This is a trait for converting strings, possibly using scientific notation, to numbers.

from_string

Implementations of FromStr and of FromStringBase, a trait for converting strings in a specified base to numbers.

to_sci

Implementations of ToSci, a trait for converting a number to string, possibly using scientific notation.

to_string

Implementations of Display, Debug, Binary, Octal, LowerHex, and UpperHex, and of the ToStringBase trait, used for converting numbers to strings.